refactor: update API tags
