+++ open-api/src/main/resources/paths/favoriteLinks.yml
@@ -1,6 +1,6 @@
 get:
   tags:
-    - externalLinks-user
+    - externalLinks
   summary: Favorite external Links List
   description: Retrieves the list of favorite external links.
   operationId: getFavoriteLinks
