--- server/src/main/java/online/dipa/hub/server/rest/TemplateController.java
@@ -1,2 +1,0 @@
-package online.dipa.hub.server.rest;public class TemplateController {
-}
