--- server/src/main/java/online/dipa/hub/services/TemplateService.java
@@ -1,2 +1,0 @@
-package online.dipa.hub.services;public class TemplateService {
-}
