+++ server/src/main/java/online/dipa/hub/services/TimelineService.java
@@ -7,6 +7,7 @@ import online.dipa.hub.api.model.*;
 import online.dipa.hub.persistence.entities.PlanTemplateEntity;
 import online.dipa.hub.persistence.entities.ProjectApproachEntity;
 import online.dipa.hub.persistence.repositories.*;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.convert.ConversionService;
 import org.springframework.stereotype.Service;
@@ -29,8 +30,8 @@ import static online.dipa.hub.api.model.Timeline.ProjectTypeEnum;
 @Transactional
 public class TimelineService {
 
-    private Map<Long, TimelineState> sessionTimelines;
-    private Map<Long, List<Template>> sessionTemplates = new HashMap<>();
+    protected static Map<Long, TimelineState> sessionTimelines = new HashMap<>();
+    protected static Map<Long, List<TimelineTemplate>> sessionTimelineTemplates = new HashMap<>();
 
     @Autowired
     private ConversionService conversionService;
@@ -42,28 +43,23 @@ public class TimelineService {
     private FileRepository fileRepository;
 
     @Autowired
-    private ProjectApproachRepository projectApproachRepository;
+    private IncrementService incrementService;
 
     @Autowired
-    private OperationTypeRepository operationTypeRepository;
+    private MilestoneService milestoneService;
 
-    @Autowired
-    private PlanTemplateRepository planTemplateRepository;
-
-    private static final long FIRST_MASTER_MILESTONE_ID = 21;
-    private static final long LAST_MASTER_MILESTONE_ID = 28;
-    private static final String CURRENT_TEMPLATE_NAME = "aktuell";
+    protected static final long FIRST_MASTER_MILESTONE_ID = 21;
 
     public List<Timeline> getTimelines() {
         initializeTimelines();
 
-        return this.sessionTimelines.values().stream().map(TimelineState::getTimeline).collect(Collectors.toList());
+        return sessionTimelines.values().stream().map(TimelineState::getTimeline).collect(Collectors.toList());
     }
 
     public Timeline getTimeline(final Long timelineId) {
         initializeTimelines();
 
-        return this.sessionTimelines.values().stream().map(TimelineState::getTimeline)
+        return sessionTimelines.values().stream().map(TimelineState::getTimeline)
                 .filter(t -> t.getId().equals(timelineId)).findFirst().orElseThrow(() -> new EntityNotFoundException(
                         String.format("Timeline with id: %1$s not found.", timelineId)));
     }
@@ -79,328 +75,28 @@ public class TimelineService {
                 });
     }
 
-    public List<OperationType> getOperationTypes() {
-
-        return operationTypeRepository.findAll().stream().map(p -> conversionService.convert(p, OperationType.class))
-                .collect(Collectors.toList());
-    }
-
-    public List<ProjectApproach> getProjectApproaches() {
-
-        return projectApproachRepository.findAll().stream()
-                .map(p -> conversionService.convert(p, ProjectApproach.class)).collect(Collectors.toList());
-    }
-
-    public List<Milestone> getMilestonesForTimeline(final Long timelineId) {
-        initializeMilestones(timelineId);
-
-        return this.sessionTimelines.get(timelineId).getMilestones();
-    }
-
-    private void initializeMilestones(final Long timelineId) {
-        TimelineState sessionTimeline = findTimelineState(timelineId);
-
-        final ProjectApproachEntity projectApproach = findProjectApproach(sessionTimeline.getTimeline().getProjectApproachId());
-
-        if (projectApproach != null && sessionTimeline.getMilestones() == null) {
-            if (projectApproach.isIterative()) {
-                initializeIncrements(timelineId);
-                sessionTimeline.setMilestones(this.loadMilestones(timelineId, 1));
-            } else {
-                sessionTimeline.setMilestones(this.loadMilestones(timelineId, 0));
-            }
-        }
-    }
-
-    private List<Milestone> loadMilestones(final Long timelineId, final int incrementCount) {
-        TimelineState sessionTimeline = findTimelineState(timelineId);
-
-        List<Milestone> milestones;
-
-        // get template Milestones
-        if (sessionTimeline.getTempIncrementMilestones() == null) {
-            milestones = getMilestonesFromRespository(timelineId);
-        } else {
-            milestones = new ArrayList<>(sessionTimeline.getTempIncrementMilestones());
-        }
-
-        // get initial milestones
-        if (incrementCount == 0) {
-
-            return milestones;
-        } else {
-
-            List<Milestone> incrementMilestones = new ArrayList<>();
-
-            // add master plan into milestones list
-            if (sessionTimeline.getMilestones() == null) {
-
-                incrementMilestones.add(milestones.get(0));
-                incrementMilestones.add(milestones.get(milestones.size() - 1));
-            } else {
-
-                List<Milestone> currentMilestones = sessionTimeline.getMilestones();
-                
-                incrementMilestones.add(currentMilestones.get(0));
-                incrementMilestones.add(currentMilestones.get(currentMilestones.size() - 1));
-
-            }
-            
-            milestones.removeIf(m -> m.getId().equals(FIRST_MASTER_MILESTONE_ID));
-            milestones.removeIf(m -> m.getId().equals(LAST_MASTER_MILESTONE_ID));
-
-            this.getValuesFromHashMap( getIncrementMilestones(milestones, timelineId,
-            incrementCount), incrementMilestones);  
-
-            return sortMilestones(incrementMilestones);
-        }
-    }
-
-    private HashMap<Increment, List<Milestone>> getIncrementMilestones(List<Milestone> tempMilestones,
-            final Long timelineId, final int incrementCount) {
-
-        HashMap<Increment, List<Milestone>> hashmapIncrementMilestones = new HashMap<>();
-
-        TimelineState sessionTimeline = findTimelineState(timelineId);
-
-        List<Increment> incrementsList = loadIncrementsTemplate(timelineId, incrementCount, tempMilestones, sessionTimeline.getMilestones());
-
-        Optional<LocalDate> firstDatePeriodOptional = tempMilestones.stream().map(Milestone::getDate).min(LocalDate::compareTo);
-        Optional<LocalDate> lastDatePeriodOptional = tempMilestones.stream().map(Milestone::getDate).max(LocalDate::compareTo);
-
-        if (firstDatePeriodOptional.isPresent() && lastDatePeriodOptional.isPresent()) {
-            LocalDate firstDatePeriod = firstDatePeriodOptional.get();
-            LocalDate lastDatePeriod = lastDatePeriodOptional.get();
-
-            long oldDaysBetween = DAYS.between(firstDatePeriod, lastDatePeriod);
-
-            long id = tempMilestones.stream().map(Milestone::getId).min(Long::compareTo).orElseThrow(() -> new EntityNotFoundException(
-                    String.format("Timeline with id: %1$s milestone has no id.", timelineId)));
-            long count = 0;
-
-            for (Increment increment : incrementsList) {
-                List<Milestone> incrementMilestones = new ArrayList<>();
-
-                long newDaysBetween = DAYS.between(increment.getStart().plusDays(14), increment.getEnd());
-                LocalDate newStartDateIncrement = increment.getStart().plusDays(14);
-
-                double factor = (double) newDaysBetween / oldDaysBetween;
-
-                for (Milestone m : tempMilestones) {
-
-                    long daysFromFirstDate = DAYS.between(firstDatePeriod, newStartDateIncrement);
-                    LocalDate newDateBeforeScale = m.getDate().plusDays(daysFromFirstDate);
-
-                    long relativePositionBeforeScale = DAYS.between(newStartDateIncrement, newDateBeforeScale);
-                    long newDateAfterScale = (long)(relativePositionBeforeScale * factor);
-                    long milestoneId = id + count;
-
-                    String milestoneName = m.getName();
-                    LocalDate milestoneDate = increment.getStart().plusDays(newDateAfterScale).plusDays(14);
-
-                    Milestone newMilestone = createMilestone(sessionTimeline, milestoneId, milestoneName, milestoneDate);
-                    incrementMilestones.add(newMilestone);
-
-                    count++;
-                }
-                hashmapIncrementMilestones.put(increment, incrementMilestones);
-
-            }
-
-        }
-        return hashmapIncrementMilestones;
-    }
-
-    private Milestone createMilestone(TimelineState sessionTimeline, long milestoneId, String milestoneName, LocalDate milestoneDate) {
-        Milestone newMilestone = new Milestone();
-        newMilestone.setId(milestoneId);
-        newMilestone.setName(milestoneName);
-        newMilestone.setDate(milestoneDate);
-
-        if (sessionTimeline.getMilestones() != null) {
-            Milestone oldMilestone = sessionTimeline.getMilestones().stream().filter(milestone -> milestone.getId().equals(milestoneId)).findFirst().orElse(null);
-
-            if (oldMilestone != null) {
-                newMilestone.setStatus(oldMilestone.getStatus());
-            } else {
-                newMilestone.setStatus(Milestone.StatusEnum.OPEN);
-            }
-        }
-        else {
-            newMilestone.setStatus(Milestone.StatusEnum.OPEN);
-        }
-
-        return newMilestone;
-    }
-
-    private List<Milestone> getMilestonesFromRespository(final Long timelineId) {
-        TimelineState sessionTimeline = findTimelineState(timelineId);
-
-        final ProjectApproachEntity projectApproach = findProjectApproach(sessionTimeline.getTimeline().getProjectApproachId());
-
-        List<Milestone> milestones = new ArrayList<>();
-
-        if (projectApproach != null) {
-
-            Long operationTypeId = projectApproach.getOperationType().getId();
-
-            Optional<PlanTemplateEntity> masterPlanTemplate = planTemplateRepository.findAll().stream()
-                .filter(template -> filterOperationType(template, operationTypeId)).findFirst();
-
-            masterPlanTemplate.ifPresent(planTemplate -> milestones.addAll(convertMilestones(planTemplate)));
-            
-
-            Optional<PlanTemplateEntity> planTemplate = planTemplateRepository.findAll().stream()
-                    .filter(template -> template.getProjectApproaches() != null)
-                    .filter(template -> filterProjectApproach(template, projectApproach.getId()))
-                    .filter(PlanTemplateEntity::getDefaultTemplate)
-                    .findFirst();
-            
-            planTemplate.ifPresent(template -> milestones.addAll(convertMilestones(template)));
-
-        }
-
-        return sortMilestones(milestones);
-    }
-
-    private List<Milestone> convertMilestones(PlanTemplateEntity planTemplate) {
-        return planTemplate.getMilestones().stream().map(m -> conversionService.convert(m, Milestone.class))
-                .sorted(Comparator.comparing(Milestone::getDate)).collect(Collectors.toList());
-    }
-
-    private List<Milestone> sortMilestones(List<Milestone> milestones) {
-         return milestones.stream().map(m -> conversionService.convert(m, Milestone.class))
-        .sorted(Comparator.comparing(Milestone::getDate)).collect(Collectors.toList());
-    }
-
-    public List<Increment> getIncrementsForTimeline(final Long timelineId) {
-
-        initializeIncrements(timelineId);
-
-        List<Increment> result = this.sessionTimelines.get(timelineId).getIncrements();
-
-        return Objects.requireNonNullElse(result, Collections.emptyList());
-    }
-
-    private void initializeIncrements(final Long timelineId) {
-        TimelineState sessionTimeline = findTimelineState(timelineId);
-        List<Template> currentSessionTemplates = this.sessionTemplates.get(timelineId);
-
-        if (sessionTimeline.getIncrements() == null) {
-            final ProjectApproachEntity projectApproach = projectRespository.findAll().stream()
-            .filter(p -> p.getProjectApproach().getId().equals(sessionTimeline.getTimeline().getProjectApproachId())).findFirst().orElseThrow(() -> new EntityNotFoundException(
-                            String.format("No project approach available for timeline id: %1$s.", timelineId))).getProjectApproach();
-
-            if (projectApproach.isIterative() && (currentSessionTemplates == null || sessionTimeline.getIterative())) {
-                
-                List<Milestone> initMilestones = getMilestonesFromRespository(timelineId);
-                sessionTimeline.setIncrements(this.loadIncrementsTemplate(timelineId, 1, initMilestones, null));
-                
-            }
-    
-        }
-    }
-
-    private List<Increment> loadIncrementsTemplate(final Long timelineId, final int incrementCount, List<Milestone> initMilestones, List<Milestone> currentMilestones) {
-
-        LocalDate firstMilestoneDate;
-        LocalDate lastMilestoneDate;
-        long daysBetween;
-        List<Increment> incrementsList = new ArrayList<>();
-
-        if (currentMilestones == null) {
-            initMilestones = sortMilestones(initMilestones);
-
-            initMilestones.removeIf(m -> m.getId().equals(FIRST_MASTER_MILESTONE_ID));
-            initMilestones.removeIf(m -> m.getId().equals(LAST_MASTER_MILESTONE_ID));
-
-            //minus 14 days for "Inkrement geplant"
-            firstMilestoneDate = initMilestones.stream().map(Milestone::getDate).min(LocalDate::compareTo).orElseThrow(() -> new EntityNotFoundException(
-                    String.format("No valid milestone date available for timeline id: %1$s.", timelineId))).minusDays(14);
-            lastMilestoneDate = initMilestones.stream().map(Milestone::getDate).max(LocalDate::compareTo).orElseThrow(() -> new EntityNotFoundException(
-                    String.format("No valid milestone date available for timeline id: %1$s.", timelineId)));
-
-        } else {
-
-            currentMilestones = sortMilestones(currentMilestones);
-
-            firstMilestoneDate = currentMilestones.get(1).getDate().minusDays(14);
-            lastMilestoneDate = currentMilestones.get(currentMilestones.size() - 2).getDate();
-
-        }
-        daysBetween = DAYS.between(firstMilestoneDate, lastMilestoneDate);
-
-        long durationIncrement = daysBetween / incrementCount;
-
-        long id = 1;
-
-        LocalDate startDateIncrement = firstMilestoneDate;
-        LocalDate endDateIncrement = startDateIncrement.plusDays(durationIncrement);
-
-        for (int i = 0; i < incrementCount; i++) {
-
-            Increment increment = new Increment();
-            increment.setId(id);
-            increment.setName("Inkrement " + id);
-            increment.setStart(startDateIncrement);
-            increment.setEnd(endDateIncrement);
-
-            incrementsList.add(increment);
-
-            id++;
-            startDateIncrement = endDateIncrement.plusDays(1);
-
-            if (i == (incrementCount - 2)) {
-                endDateIncrement = lastMilestoneDate;
-            } else {
-                endDateIncrement = endDateIncrement.plusDays(durationIncrement);
-            }
-        }
-        return incrementsList;
-    }
-
-    public void addIncrement(Long timelineId) {
-        TimelineState sessionTimeline = findTimelineState(timelineId);
-
-        int incrementCount = sessionTimeline.getIncrements().size();
-
-        sessionTimeline.setIncrements(this.loadIncrementsTemplate(timelineId, incrementCount + 1, null, sessionTimeline.getMilestones()));
-        sessionTimeline.setMilestones(this.loadMilestones(timelineId, incrementCount + 1));
-    }
-
-    public void deleteIncrement(Long timelineId) {
-        TimelineState sessionTimeline = findTimelineState(timelineId);
-
-        int incrementCount = sessionTimeline.getIncrements().size();
-
-        if (incrementCount != 1) {
-            sessionTimeline.setIncrements(this.loadIncrementsTemplate(timelineId, incrementCount - 1, null, sessionTimeline.getMilestones()));
-            sessionTimeline.setMilestones(this.loadMilestones(timelineId, incrementCount - 1));
-        }
-    }
-
-    private ProjectApproachEntity findProjectApproach(Long projectApproachId) {
+    ProjectApproachEntity findProjectApproach(Long projectApproachId) {
         return projectRespository.findAll().stream()
         .filter(p -> p.getProjectApproach().getId().equals(projectApproachId)).findFirst().orElseThrow(() -> new EntityNotFoundException(
                         String.format("Project approach with id: %1$s not found.", projectApproachId))).getProjectApproach();
     }
 
-    private TimelineState findTimelineState(Long timelineId) {
-        return getSessionTimelines().computeIfAbsent(timelineId, t -> new TimelineState());
+    public TimelineState findTimelineState(Long timelineId) {
+        return sessionTimelines.computeIfAbsent(timelineId, t -> new TimelineState());
     }
 
-    private Map<Long, TimelineState> getSessionTimelines() {
-        if (this.sessionTimelines == null) {
-            this.sessionTimelines = new HashMap<>();
+    Map<Long, TimelineState> getSessionTimelines() {
+        if (sessionTimelines == null) {
+            sessionTimelines = new HashMap<>();
         }
-        return this.sessionTimelines;
+        return sessionTimelines;
     }
 
     public void moveTimelineByDays(final Long timelineId, final Long days) {
 
         TimelineState sessionTimeline = getSessionTimelines().get(timelineId);
 
-        updateTempMilestones(timelineId);
+        milestoneService.updateTempMilestones(timelineId);
 
         LocalDate newTimelineStart = sessionTimeline.getTimeline().getStart().plusDays(days);
         LocalDate newTimelineEnd = sessionTimeline.getTimeline().getEnd().plusDays(days);
@@ -416,7 +112,7 @@ public class TimelineService {
             temp.setDate(temp.getDate().plusDays(days));
         }
 
-        updateIncrements(timelineId);
+        incrementService.updateIncrements(timelineId);
 
     }
 
@@ -424,7 +120,7 @@ public class TimelineService {
 
         TimelineState sessionTimeline = getSessionTimelines().get(timelineId);
 
-        updateTempMilestones(timelineId);
+        milestoneService.updateTempMilestones(timelineId);
 
         LocalDate timelineStart = sessionTimeline.getTimeline().getStart();
         LocalDate timelineEnd = sessionTimeline.getTimeline().getEnd();
@@ -441,7 +137,7 @@ public class TimelineService {
             long newMilestoneRelativePosition = Math.round(oldMilestoneRelativePosition * factor);
 
             m.setDate(newTimelineStart.plusDays(newMilestoneRelativePosition));
-           
+
         }
 
         for (Milestone temp : sessionTimeline.getTempIncrementMilestones()) {
@@ -449,10 +145,10 @@ public class TimelineService {
             long newMilestoneRelativePosition = Math.round(oldMilestoneRelativePosition * factor);
 
             temp.setDate(newTimelineStart.plusDays(newMilestoneRelativePosition));
-            
+
         }
 
-        updateIncrements(timelineId);
+        incrementService.updateIncrements(timelineId);
 
     }
 
@@ -460,7 +156,7 @@ public class TimelineService {
 
         TimelineState sessionTimeline = getSessionTimelines().get(timelineId);
 
-        updateTempMilestones(timelineId);
+        milestoneService.updateTempMilestones(timelineId);
 
         LocalDate timelineStart = sessionTimeline.getTimeline().getStart();
         LocalDate timelineEnd = sessionTimeline.getTimeline().getEnd();
@@ -486,7 +182,7 @@ public class TimelineService {
             temp.setDate(timelineStart.plusDays(newMilestoneRelativePosition));
         }
 
-        updateIncrements(timelineId);
+        incrementService.updateIncrements(timelineId);
 
     }
 
@@ -550,7 +246,7 @@ public class TimelineService {
             String projectEventTitle = "Projektstart" + " - " + projectApproach.getName();
             icsCalendar.addEvent(timezone, sessionTimeline.getTimeline().getStart(), projectEventTitle, "Test Comment");
 
-            List<Milestone> milestones = getMilestonesForTimeline(timelineId);
+            List<Milestone> milestones = milestoneService.getMilestonesForTimeline(timelineId);
             for (Milestone milestone : milestones) {
                 LocalDate eventDate = milestone.getDate();
                 String eventTitle = milestone.getName() + " - " + projectApproach.getName();
@@ -563,31 +259,6 @@ public class TimelineService {
         return icsCalendar.getCalendarFile("Meilensteine");
     }
 
-    private void updateIncrements(final Long timelineId) {
-
-        TimelineState sessionTimeline = getSessionTimelines().get(timelineId);
-
-        List<Increment> increments = sessionTimeline.getIncrements();
-
-        if (increments != null) {
-            int incrementCount = increments.size();
-            sessionTimeline.setIncrements(this.loadIncrementsTemplate(timelineId, incrementCount, null, sessionTimeline.getMilestones()));
-            // fix sessionTimeline setMilestones
-
-        }
-    }
-
-    private void updateTempMilestones(final Long timelineId) {
-
-        TimelineState sessionTimeline = getSessionTimelines().get(timelineId);
-
-        if (sessionTimeline.getTempIncrementMilestones() == null) {
-
-            List<Milestone> tempMilestonesList = new ArrayList<>(getMilestonesFromRespository(timelineId));
-            sessionTimeline.setTempIncrementMilestones(tempMilestonesList);
-        }
-    }
-
     public void updateTimeline(final Timeline timeline) {
 
         TimelineState sessionTimeline = getSessionTimelines().get(timeline.getId());
@@ -598,173 +269,10 @@ public class TimelineService {
         sessionTimeline.setIncrements(null);
         sessionTimeline.setTempIncrementMilestones(null);
 
-        this.sessionTemplates.remove(timeline.getId());
+        sessionTimelineTemplates.remove(timeline.getId());
 
-        this.initializeMilestones(timeline.getId());
-        this.updateMilestonesAndIncrement(timeline);
-    }
-
-    public void updateMilestonesAndIncrement(final Timeline timeline) {
-
-        TimelineState sessionTimeline = getSessionTimelines().get(timeline.getId());
-
-        updateTempMilestones(timeline.getId());
-
-        LocalDate initTimelineStart = LocalDate.now();
-        Optional<LocalDate> initTimelineEndOptional = sessionTimeline.getMilestones().stream().map(Milestone::getDate).max(LocalDate::compareTo);
-
-        if (initTimelineEndOptional.isPresent()) {
-            LocalDate initTimelineEnd = initTimelineEndOptional.get();
-
-            LocalDate currentTimelineStart = sessionTimeline.getTimeline().getStart();
-            LocalDate currentTimelineEnd = sessionTimeline.getTimeline().getEnd();
-
-            long oldDaysBetween = DAYS.between(initTimelineStart, initTimelineEnd);
-            long newDaysBetween = DAYS.between(currentTimelineStart, currentTimelineEnd);
-            double factor = (double) newDaysBetween / oldDaysBetween;
-
-            for (Milestone m : sessionTimeline.getMilestones()) {
-
-                long oldMilestoneRelativePosition = DAYS.between(initTimelineStart, m.getDate());
-                long newMilestoneRelativePosition = Math.round(oldMilestoneRelativePosition * factor);
-
-                m.setDate(currentTimelineStart.plusDays(newMilestoneRelativePosition));          
-                
-            }
-
-            for (Milestone temp : sessionTimeline.getTempIncrementMilestones()) {
-                long oldMilestoneRelativePosition = DAYS.between(initTimelineStart, temp.getDate());
-                long newMilestoneRelativePosition = Math.round(oldMilestoneRelativePosition * factor);
-
-                temp.setDate(currentTimelineStart.plusDays(newMilestoneRelativePosition));
-                
-            }
-
-            updateIncrements(timeline.getId());
-        }
-    }
-
-    public void updateMilestoneStatus(final Long timelineId, final Long milestoneId, final Milestone.StatusEnum status) {
-        final TimelineState sessionTimeline = getSessionTimelines().get(timelineId);
-        final Milestone updatedMilestone = sessionTimeline.getMilestones().stream().filter(m -> m.getId().equals(milestoneId)).findFirst().orElseThrow(() -> new EntityNotFoundException(
-                String.format("Milestone with id: %1$s not found.", milestoneId)));
-
-        updatedMilestone.setStatus(status);
-    }
-
-    public List<Template> getTemplatesForTimeline(final Long timelineId) {
-
-        List<Template> templates = new ArrayList<>();
-        final TimelineState sessionTimeline = getSessionTimelines().get(timelineId);
-        long count = 0;
-
-        Template currentTemplate = new Template()
-                                        .id(count++)
-                                        .name(CURRENT_TEMPLATE_NAME)
-                                        .milestones(getMilestonesForTimeline(timelineId))
-                                        .increments(getIncrementsForTimeline(timelineId));
-
-        templates.add(currentTemplate);
-
-        final ProjectApproachEntity projectApproach = findProjectApproach(sessionTimeline.getTimeline().getProjectApproachId());
-        Long operationTypeId = projectApproach.getOperationType().getId();
-
-        Optional<PlanTemplateEntity> masterPlanTemplate = planTemplateRepository.findAll().stream()
-                                                            .filter(template -> filterOperationType(template, operationTypeId))
-                                                            .findFirst();
-
-        List<PlanTemplateEntity> projectApproachPlanTemplates = planTemplateRepository.findAll().stream()
-                .filter(template -> template.getProjectApproaches() != null)
-                .filter(template -> filterProjectApproach(template, projectApproach.getId()))
-                .collect(Collectors.toList());
-
-        for (PlanTemplateEntity temp: projectApproachPlanTemplates) {
-            List<Milestone> milestones = new ArrayList<>();
-
-            if (masterPlanTemplate.isPresent()) {
-                milestones.addAll(convertMilestones(masterPlanTemplate.get()));
-            }
-
-            Template template = new Template()
-                                .id(count++)
-                                .name(temp.getName())
-                                .standard(temp.getStandard());
-
-            if (projectApproach.isIterative()) {
-
-                milestones = updateMilestonesTemplate(timelineId, milestones);
-                List<Milestone> tempMilestones = new ArrayList<>(convertMilestones(temp));
-
-                this.getValuesFromHashMap(getIncrementMilestones(tempMilestones,timelineId, 1), milestones);
-
-                template.milestones(sortMilestones(milestones));
-                
-                if (temp.getStandard()) {
-                    template.increments(loadIncrementsTemplate((long) 1, 1, milestones, null));
-                }
-
-            }
-            else {
-                milestones.addAll(convertMilestones(temp));
-                template.milestones(updateMilestonesTemplate(timelineId, milestones));
-            }
-
-            templates.add(template);
-
-        }
-
-        this.sessionTemplates.put(timelineId, templates);
-        return templates;
-      
-    }
-
-    public List<Milestone> updateMilestonesTemplate(final Long timelineId, List<Milestone> milestones) {
-
-        TimelineState sessionTimeline = getSessionTimelines().get(timelineId);
-        milestones = sortMilestones(milestones);
-
-        LocalDate initTimelineStart = LocalDate.now();
-
-        Optional<LocalDate> initTimelineEndOptional = milestones.stream().map(Milestone::getDate).max(LocalDate::compareTo);
-
-        LocalDate initTimelineEnd;
-
-        if (initTimelineEndOptional.isPresent()) {
-            initTimelineEnd = initTimelineEndOptional.get();
-
-            LocalDate currentTimelineStart = sessionTimeline.getTimeline().getStart();
-            LocalDate currentTimelineEnd = sessionTimeline.getTimeline().getEnd();
-
-            long oldDaysBetween = DAYS.between(initTimelineStart, initTimelineEnd);
-            long newDaysBetween = DAYS.between(currentTimelineStart, currentTimelineEnd);
-            double factor = (double) newDaysBetween / oldDaysBetween;
-
-            for (Milestone m : milestones) {
-                long oldMilestoneRelativePosition = DAYS.between(initTimelineStart, m.getDate());
-                long newMilestoneRelativePosition = Math.round(oldMilestoneRelativePosition * factor);
-
-                m.setDate(currentTimelineStart.plusDays(newMilestoneRelativePosition));
-             
-            }
-        }
-
-        return milestones;
-    }
-
-    public void updateTemplateForProject(final Long timelineId, final Long templateId) {
-        TimelineState sessionTimeline = getSessionTimelines().get(timelineId);
-
-        List<Template> currentSessionTemplates = this.sessionTemplates.get(timelineId);
-        Optional<Template> selectedTemplateOptional = currentSessionTemplates.stream().filter(t -> t.getId().equals(templateId)).findFirst();
-
-        if (selectedTemplateOptional.isPresent()) {
-            Template selectedTemplate = selectedTemplateOptional.get();
-            sessionTimeline.setMilestones(selectedTemplate.getMilestones());
-            sessionTimeline.setIncrements(selectedTemplate.getIncrements());
-            sessionTimeline.setIterative(selectedTemplate.getIncrements() != null);
-
-        }
-
+        milestoneService.initializeMilestones(timeline.getId());
+        milestoneService.updateMilestonesAndIncrement(timeline);
     }
 
     public List<DownloadFile> getFilesForMilestone(final Long timelineId, final Long milestoneId) {
@@ -809,28 +317,28 @@ public class TimelineService {
         return downloadFileIds;
     }
 
-    private void getValuesFromHashMap(HashMap<Increment, List<Milestone>> hashMap, List<Milestone> list) {
-        
+    void getValuesFromHashMap(HashMap<Increment, List<Milestone>> hashMap, List<Milestone> list) {
+
         for (List<Milestone> value : hashMap.values()) {
             list.addAll(value);
         }
     }
 
-    
-    private boolean filterOperationType(PlanTemplateEntity template, final Long operationTypeId) {
+
+    boolean filterOperationType(PlanTemplateEntity template, final Long operationTypeId) {
         Optional<OperationType> operationType = template.getOperationTypes().stream()
             .map(p -> conversionService.convert(p, OperationType.class))
             .filter(o -> o.getId().equals(operationTypeId)).findFirst();
-        
+
         return operationType.isPresent();
 
     }
 
-    private boolean filterProjectApproach(PlanTemplateEntity template, final Long projectApproachId) {
+    boolean filterProjectApproach(PlanTemplateEntity template, final Long projectApproachId) {
         Optional<ProjectApproach> projectApproach = template.getProjectApproaches().stream()
             .map(p -> conversionService.convert(p, ProjectApproach.class))
             .filter(o -> o.getId().equals(projectApproachId)).findFirst();
-        
+
         return projectApproach.isPresent();
 
     }
